requirements.txt
dask[delayed]
natcap.invest==3.8.0
numpy
rasterio
scikit-learn
simanneal
